experiment with OSM street network accelerator
The original batch generator took about 544 minutes (9 hours + 4 minutes) to download all of the "highway" networks in the agglomerations of Togo, Ghana, Tanzania, and Kenya. This was the third and final attempt to accelerate the process. The idea was to adopt the architecture used in Google's building footprint downloader.
It took this version 622 minutes (10 hours 22 minutes)